Move files and change namespaces
