+++ src/Razor/src/Microsoft.CodeAnalysis.Razor.Workspaces/FoldingRanges/IRazorFoldingRangeProvider.cs
@@ -7,7 +7,7 @@ using System.Threading.Tasks;
 using Microsoft.CodeAnalysis.Razor.ProjectSystem;
 using Microsoft.VisualStudio.LanguageServer.Protocol;
 
-namespace Microsoft.AspNetCore.Razor.LanguageServer.Folding;
+namespace Microsoft.CodeAnalysis.Razor.FoldingRanges;
 
 internal interface IRazorFoldingRangeProvider
 {
