+++ src/Razor/src/Microsoft.CodeAnalysis.Razor.Workspaces/FoldingRanges/SectionDirectiveFoldingProvider.cs
@@ -6,7 +6,7 @@ using Microsoft.AspNetCore.Razor.Language;
 using Microsoft.AspNetCore.Razor.Language.Syntax;
 using Microsoft.CodeAnalysis.Razor.Workspaces;
 
-namespace Microsoft.AspNetCore.Razor.LanguageServer.Folding;
+namespace Microsoft.CodeAnalysis.Razor.FoldingRanges;
 
 internal sealed class SectionDirectiveFoldingProvider : AbstractSyntaxNodeFoldingProvider<RazorDirectiveSyntax>
 {
